camber_bundle/worker/requirements.txt
pytesseract
Pillow
requests